Move the object store call tag into its own module
