+++ slatedb/src/lib.rs
@@ -140,6 +140,7 @@ mod mem_table;
 mod memtable_flusher;
 mod merge_iterator;
 mod merge_operator;
+mod object_store_tag;
 mod object_stores;
 mod ops;
 mod oracle;
